allow csharp driven modding

+++ scripts/modding/SpriteFramesBuilder.cs
@@ -2,6 +2,9 @@ using Godot;
 
 namespace OmoriSandbox.Modding;
 
+/// <summary>
+/// A helper class to build SpriteFrames from a texture atlas.
+/// </summary>
 public class SpriteFramesBuilder
 {
     private SpriteFrames spriteFrames;
@@ -9,6 +12,14 @@ public class SpriteFramesBuilder
     private int Width;
     private int Height;
     private int Columns;
+
+    /// <summary>
+    /// Creates a new SpriteFramesBuilder.<br/>You can call <see cref="AddEmotion(string, double, int[])"/> to different emotions to the list of animations.
+    /// </summary>
+    /// <param name="atlasPath">The path to the atlas file. Must be a full path from the root mods folder.<br/>
+    /// Example: <c>MyMod/actors/MyActor/atlas.png</c></param>
+    /// <param name="width">The width of a single sprite in the atlas.</param>
+    /// <param name="height">The height of a single sprite in the atlas.</param>
     public SpriteFramesBuilder(string atlasPath, int width, int height)
     {
         if (!FileAccess.FileExists("user://mods/" + atlasPath))
@@ -23,7 +34,14 @@ public class SpriteFramesBuilder
         Columns = texture.GetWidth() / Width;
     }
 
-    public SpriteFramesBuilder AddEmotion(string emotion, int fps, params int[] indices)
+    /// <summary>
+    /// Adds an emotion animation to the current SpriteFramesBuilder.<br/>
+    /// </summary>
+    /// <param name="emotion">The emotion this animation corresponds to.</param>
+    /// <param name="fps">The FPS of the animation.</param>
+    /// <param name="indices">A list of indices into the atlas. Index 0 would be the top left of your altas, and increments going left to right.</param>
+    /// <returns></returns>
+    public SpriteFramesBuilder AddEmotion(string emotion, double fps, params int[] indices)
     {
         if (spriteFrames.HasAnimation(emotion))
         {
@@ -48,6 +66,7 @@ public class SpriteFramesBuilder
         return this;
     }
 
+    /// <returns>The built <see cref="SpriteFrames"/> object to use.</returns>
     public SpriteFrames Build()
     {
         return spriteFrames;
